backend/requirements.txt
fastapi
uvicorn
openai
transformers
torch
pillow
langchain
faiss-cpu
python-multipart
pymupdf  # PDF reader
sentence-transformers
python-docx
redis
